Refactored the code: it works!
