--- infra/deploy/sync-request.json
@@ -1,11 +1,0 @@
-{
-    "config": {
-        "encoding":"FLAC",
-        "sampleRateHertz": 16000,
-        "languageCode": "fr-FR",
-        "enableWordTimeOffsets": false
-    },
-    "audio": {
-        "uri":"gs://cloud-samples-tests/speech/brooklyn.flac"
-    }
-}
