--- infra/deploy/terraform.tfvars
@@ -1,2 +1,0 @@
-# PROJECT COMPONENTS
-aws_extractor = "audio-extractor"
